section_2/requirements.txt
Pillow~=9.2.0
pip~=22.0.4
wheel~=0.37.1
mysql-connector-python~=8.0.30
rsa~=4.9
pyasn1~=0.4.8
gast~=0.4.0
six~=1.16.0
numpy~=1.23.1
h5py~=3.7.0
keras~=2.9.0
tensorflow~=2.9.1
wrapt~=1.14.1
requests~=2.28.1
setuptools~=58.1.0
cachetools~=5.2.0
certifi~=2022.6.15
urllib3~=1.26.11
python-dateutil~=2.8.2
Markdown~=3.4.1
oauthlib~=3.2.0
matplotlib~=3.5.2
idna~=3.3
MarkupSafe~=2.1.1
Werkzeug~=2.2.1
fonttools~=4.34.4
pyparsing~=3.0.9
cycler~=0.11.0
packaging~=21.3
kiwisolver~=1.4.4
tensorboard~=2.9.1
flatbuffers~=1.12
astunparse~=1.6.3
termcolor~=1.1.0
zipp~=3.8.1
py4j~=0.10.9.5
protobuf~=3.19.4
pyspark~=3.3.0
